Created search command
